Update to publish toolbox overview
This change updates the toolbox overview to 1.2 and also updates the release for the fingerprint toolbox.

Moved the toolboxes into a new table separate from the main docs and a heading for it. Also added date columns to the tables marking the date of the publication of the doc/toolbox.